Cross-Animal + Dev x Null Updates
Cross-Animal:
- functions/compare_conditions_analysis.py: reorg w/ updated function calls
- functions/compare_datasets_funcs.py: remove functions to move to own files.
- functions/cross_animal_dev_null_plots.py: all dev x null plotting code moved here
- functions/cross_animal_dev_stats.py: all dev stat code moved here

Dev x Null updates:
- functions/dev_null_cutoff_dicts.py: move code to calculate dev x null dicts here
- functions/dev_plot_funcs.py: plot close added
- functions/deviation_null_analysis.py: updated to have firing rate dictionaries calculated and stored
