+++ functions/compare_datasets_funcs.py
@@ -3450,310 +3450,6 @@ def cross_dataset_pop_rate_taste_corr_plots(rate_corr_data, unique_given_names,
         f_taste.savefig(os.path.join(results_dir,corr_type + '_by_taste_mean_trends.png'))
         f_taste.savefig(os.path.join(results_dir,corr_type + '_by_taste_mean_trends.svg'))
         plt.close(f_taste)
-            
-def cross_dataset_dev_stats_plots(dev_stats_data, unique_given_names, 
-                                            unique_dev_stats_names, unique_segment_names, 
-                                            results_dir):   
-    """This function is dedicated to plotting deviation statistics across animals.
-    INPUTS:
-        - dev_stats_data: dictionary containing data regarding deviation statistics
-            organized as follows:
-                - dev_stats_data.keys() are the unique_given_names
-                - dev_stats_data[name]['dev_stats'] = dict containing deviation stats
-                - dev_stats_data[name]['dev_stats'].keys() are the unique_dev_stats_names
-                - dev_stats_data[name]['dev_stats'][stats_name] = dict containing specific statistic results
-                - dev_stats_data[name]['dev_stats'][stats_name].keys() are the unique_segment_names
-                - dev_stats_data[name]['dev_stats'][stats_name][seg_name] = array of deviation statistics
-        - unique_given_names: names of imported datasets in dict
-        - unique_dev_stats_names: names of types of statistics
-        - unique_segment_names: names of segments analyzed
-        - results_dir: storage directory
-    OUTPUTS:
-        - Plots with statistical results
-    """
-    colors = ['green','royalblue','blueviolet','teal','deeppink', \
-              'springgreen','turquoise', 'midnightblue', 'lightskyblue', \
-              'palevioletred', 'darkslateblue']
-    
-    plot_side = np.ceil(np.sqrt(len(unique_dev_stats_names))).astype('int')
-    plot_inds = np.reshape(np.arange(plot_side**2),(plot_side,plot_side))
-    seg_pairs = list(combinations(np.arange(len(unique_segment_names)),2))
-    f_stats, ax_stats = plt.subplots(nrows=plot_side,ncols=plot_side,
-                                     figsize=(8,8),sharex=True)
-    for ds_i, ds in enumerate(unique_dev_stats_names):
-        ds_name = (' ').join(ds.split('_')[:-1])
-        ax_r, ax_c = np.where(plot_inds == ds_i)
-        all_animal_stats = np.zeros((len(unique_given_names),len(unique_segment_names)))
-        for gn_i, gn in enumerate(unique_given_names):
-            seg_means = np.zeros(len(unique_segment_names))
-            for s_i, sn in enumerate(unique_segment_names):
-                all_animal_stats[gn_i,s_i] = np.nanmean(dev_stats_data[gn]['dev_stats'][ds][sn])
-        ax_stats[ax_r[0],ax_c[0]].boxplot(all_animal_stats)
-        for s_i in range(len(unique_segment_names)):
-            scat_x = s_i+1+np.random.randn(len(unique_given_names))/10
-            ax_stats[ax_r[0],ax_c[0]].scatter(scat_x,all_animal_stats[:,s_i],\
-                                              color='g',alpha=0.3)
-        ax_stats[ax_r[0],ax_c[0]].set_xticks(np.arange(len(unique_segment_names))+1,unique_segment_names,\
-                                 rotation=45)
-        ax_stats[ax_r[0],ax_c[0]].set_ylabel(ds_name)
-        max_y = np.nanmax(all_animal_stats)
-        #ANOVA test
-        result = f_oneway(*list(all_animal_stats.T))
-        if result.pvalue <= 0.05:
-            ax_stats[ax_r[0],ax_c[0]].set_title(ds_name + ' *ANOVA')
-        else:
-            ax_stats[ax_r[0],ax_c[0]].set_title(ds_name)
-        #Pairwise TTest
-        for sp_0, sp_1 in seg_pairs:
-            result = ttest_ind(all_animal_stats[:,sp_0],all_animal_stats[:,sp_1])
-            if result.pvalue <= 0.05:
-                max_y += max_y*0.1
-                plt.plot([sp_0+1,sp_1+1],[max_y,max_y])
-                max_y += max_y*0.1
-                plt.text(1+sp_0+(sp_1-sp_0)/2,max_y,'*TT')
-            result = ks_2samp(all_animal_stats[:,sp_0],all_animal_stats[:,sp_1])
-            if result.pvalue <= 0.05:
-                max_y += max_y*0.1
-                plt.plot([sp_0+1,sp_1+1],[max_y,max_y])
-                max_y += max_y*0.1
-                plt.text(1+sp_0+(sp_1-sp_0)/2,max_y,'*KS')
-    plt.tight_layout()
-    f_stats.savefig(os.path.join(results_dir,'overview_boxplots.png'))
-    f_stats.savefig(os.path.join(results_dir,'overview_boxplots.svg'))
-  
-    #By-stat distributions
-    for dev_stat in unique_dev_stats_names:
-        max_val = 0
-        max_mean = 0
-        cross_animal_means = []
-        combined_animal_results = []
-        for seg_name in unique_segment_names:
-            combined_seg_results = []
-            seg_animal_means = []
-            for name in unique_given_names:
-                dataset = dev_stats_data[name]['dev_stats'][dev_stat][seg_name]
-                try:
-                    num_vals = len(dataset)
-                    combined_seg_results.extend(list(dataset))
-                    seg_animal_means.extend([np.nanmean(dataset)])
-                except:
-                    if (dataset.dtype == float) or (dataset.dtype == int):
-                        combined_seg_results.extend([dataset])
-                        seg_animal_means.extend([dataset])
-            if max(np.array(combined_seg_results)) > max_val:
-                max_val = max(np.array(combined_seg_results))
-            if max(np.array(seg_animal_means)) > max_mean:
-                max_mean = max(np.array(seg_animal_means))
-            cross_animal_means.append(seg_animal_means)
-            combined_animal_results.append(combined_seg_results)
-        max_val = np.ceil(max_val).astype('int')
-        max_mean = np.ceil(max_mean).astype('int')
-        #Create figure of combined data
-        f, ax = plt.subplots(nrows=1, ncols=2, figsize = (4*2,4))
-        #plt.boxplot(combined_animal_results,labels=unique_segment_names)
-        for s_i, seg_name in enumerate(unique_segment_names):
-            cdf_data = np.array(combined_animal_results[s_i])
-            nonnan_data = cdf_data[~np.isnan(cdf_data)]
-            min_x = np.nanmin(nonnan_data)
-            max_x = np.nanmax(nonnan_data)
-            cdf_x = np.linspace(min_x,max_x,1000)
-            cdf_vals = np.array([len(np.where(nonnan_data <= i)[0])/len(nonnan_data) for i in cdf_x])
-            ax[0].plot(cdf_x,cdf_vals,label=seg_name,color=colors[s_i])
-            ax[1].hist(nonnan_data,bins=max_val,histtype='step',\
-                     density=True,label='seg_name',color=colors[s_i])
-            data_mean = np.nanmean(combined_animal_results[s_i])
-            ax[1].axvline(data_mean,color=colors[s_i],\
-                          label=seg_name + 'mean = ' + str(np.round(data_mean,2)))
-        ax[0].legend(loc='upper left')
-        ax[1].legend(loc='upper right')
-        ax[0].set_xlabel(dev_stat)
-        ax[1].set_xlabel(dev_stat)
-        ax[0].set_ylabel('Cumulative Fraction')
-        ax[1].set_xlabel('Density')
-        ax[0].set_title('Cumulative Distribution')
-        ax[1].set_title('Density Distribution')
-        seg_pairs = list(combinations(np.arange(len(unique_segment_names)),2))
-        pair_statistics = ('Segment 1').ljust(10,' ') + ' | ' + ('Segment 2').ljust(10,' ') + ' | ' + ('TTest').ljust(5,' ')
-        for sp in seg_pairs:
-            seg_1 = sp[0]
-            seg_1_data = np.array(combined_animal_results[seg_1])
-            seg_2 = sp[1]
-            seg_2_data = np.array(combined_animal_results[seg_2])
-            ttest_result = ttest_ind(seg_1_data[~np.isnan(seg_1_data)],seg_2_data[~np.isnan(seg_2_data)])
-            if ttest_result[1] <= 0.05:
-                pair_statistics = pair_statistics + '\n' + \
-                    unique_segment_names[seg_1].ljust(10, ' ') + '   ' + \
-                        unique_segment_names[seg_2].ljust(10, ' ') + '   ' + ('*').ljust(5,' ')
-            else:
-                pair_statistics = pair_statistics + '\n' + \
-                    unique_segment_names[seg_1].ljust(10, ' ') + '   ' + \
-                        unique_segment_names[seg_2].ljust(10, ' ') + '   ' + ('n.s.').ljust(5,' ')
-        ax[0].text(max_val,0,pair_statistics,horizontalalignment='right',\
-                 verticalalignment='bottom',fontsize=10,color='k',
-                 bbox=dict(boxstyle="round",color="grey", alpha=0.5))
-        plt.tight_layout()
-        f.savefig(os.path.join(results_dir,dev_stat + '_distributions.png'))
-        f.savefig(os.path.join(results_dir,dev_stat + '_distributions.svg'))
-        plt.close(f)
-        
-        #Create figure of animal means
-        f_means = plt.figure(figsize=(5,5))
-        for n_i, name in enumerate(unique_given_names):
-            animal_means = []
-            for s_i, seg_name in enumerate(unique_segment_names):
-                animal_means.extend([cross_animal_means[s_i][n_i]])
-            plt.plot(np.arange(len(unique_segment_names)),animal_means,\
-                     color=colors[n_i],alpha=0.5,label=name + ' Mean')
-        all_animal_means = []
-        all_animal_stds = []
-        for s_i, seg_name in enumerate(unique_segment_names):
-            all_animal_means.extend([np.nanmean(combined_animal_results[s_i])])
-            all_animal_stds.extend([np.nanstd(combined_animal_results[s_i])])
-        plt.plot(np.arange(len(unique_segment_names)),all_animal_means,\
-                 color='k',linestyle='dashed',label='Mean')
-        plt.fill_between(np.arange(len(unique_segment_names)),\
-                         np.array(all_animal_means) - np.array(all_animal_stds),\
-                        np.array(all_animal_means) + np.array(all_animal_stds),\
-                        color='k',alpha=0.2,label='Std')
-        plt.xticks(np.arange(len(unique_segment_names)),unique_segment_names)
-        plt.legend()
-        plt.ylim([0,max_mean + np.nanmax(all_animal_stds)])
-        plt.title(dev_stat)
-        f_means.savefig(os.path.join(results_dir,dev_stat + '_means.png'))
-        f_means.savefig(os.path.join(results_dir,dev_stat + '_means.svg'))
-        plt.close(f_means)
-        
-def cross_dataset_dev_null_plots(dev_null_data, unique_given_names, 
-                                 unique_dev_null_names, unique_segment_names, 
-                                 results_dir):   
-    """This function is dedicated to plotting deviation statistics compared
-    to null distributions across animals.
-    INPUTS:
-        - dev_null_data: dictionary containing data regarding deviation statistics
-            organized as follows:
-                - dev_null_data.keys() are the unique_given_names
-                - dev_null_data[name]['dev_null'] = dict containing deviation stats
-                - dev_null_data[name]['dev_null'].keys() are the unique_dev_stats_names
-                - dev_null_data[name]['dev_null'][null_name] = dict containing specific statistic results
-                - dev_null_data[name]['dev_null'][null_name].keys() are the unique_segment_names
-                - dev_null_data[name]['dev_null'][null_name][seg_name].keys() = ['null','true','percentile']
-                - dev_null_data[name]['dev_null'][null_name][seg_name]['null'] = list of [[cutoff value],[mean count],[std]]
-                - dev_null_data[name]['dev_null'][null_name][seg_name]['true'] = list of [[cutoff value],[count]]
-        - unique_given_names: names of imported datasets in dict
-        - unique_dev_null_names: names of types of statistics
-        - unique_segment_names: names of segments analyzed
-        - results_dir: storage directory
-    OUTPUTS:
-        - Plots with statistical results
-    """
-    colors = ['green','royalblue','blueviolet','teal','deeppink', \
-              'springgreen','turquoise', 'midnightblue', 'lightskyblue', \
-              'palevioletred', 'darkslateblue']
-    for dev_null_stat in unique_dev_null_names:
-        max_val = 0
-        f, ax = plt.subplots(ncols=len(unique_segment_names),figsize=(4*len(unique_segment_names),4))
-        f_log, ax_log = plt.subplots(ncols=len(unique_segment_names),figsize=(4*len(unique_segment_names),4))
-        for s_i,seg_name in enumerate(unique_segment_names):
-            #Collect across animals data
-            max_cutoff_val = 0
-            min_cutoff_val = 100000
-            combined_cutoff_values = []
-            seg_animal_true_vals = []
-            seg_animal_null_means = []
-            seg_animal_null_stds = []
-            for name in unique_given_names:
-                null_dataset = dev_null_data[name]['dev_null'][dev_null_stat][seg_name]['null']
-                true_dataset = dev_null_data[name]['dev_null'][dev_null_stat][seg_name]['true']
-                combined_cutoff_values.append(null_dataset[0])
-                seg_animal_null_means.append(null_dataset[1])
-                seg_animal_null_stds.append(null_dataset[2])
-                seg_animal_true_vals.append(true_dataset[1])
-                if max(np.array(true_dataset[1])) > max_val:
-                    max_val = max(np.array(true_dataset[1]))
-                if max(np.array(null_dataset[1])) + max(np.array(null_dataset[2])) > max_val:
-                    max_val = max(np.array(null_dataset[1])) + max(np.array(null_dataset[2]))
-                if max(np.array(null_dataset[0])) > max_cutoff_val:
-                    max_cutoff_val = max(np.array(null_dataset[0]))
-                if min(np.array(null_dataset[0])) < min_cutoff_val:
-                    min_cutoff_val = min(np.array(null_dataset[0]))
-            #Reorganize data
-            cutoff_array = np.arange(min_cutoff_val,max_cutoff_val+1)
-            num_anim = len(seg_animal_true_vals)
-            num_x_vals = len(cutoff_array)
-            true_array = np.nan*np.ones((num_anim,num_x_vals))
-            null_mean_array = np.nan*np.ones((num_anim,num_x_vals))
-            null_std_array = np.nan*np.ones((num_anim,num_x_vals))
-            for a_i in range(num_anim):
-                anim_cutoffs = combined_cutoff_values[a_i]
-                anim_true_means = seg_animal_true_vals[a_i]
-                anim_null_means = seg_animal_null_means[a_i]
-                anim_null_stds = seg_animal_null_stds[a_i]
-                for c_i, c_val in enumerate(anim_cutoffs):
-                    c_array_ind = np.where(cutoff_array == c_val)[0]
-                    true_array[a_i,c_array_ind] = anim_true_means[c_i]
-                    null_mean_array[a_i,c_array_ind] = anim_null_means[c_i]
-                    null_std_array[a_i,c_array_ind] = anim_null_stds[c_i]
-            #Generate null data from mean and std data
-            new_null_mean_array = np.nan*np.ones(len(cutoff_array))
-            new_null_std_array = np.nan*np.ones(len(cutoff_array))
-            for c_i,c_val in enumerate(cutoff_array):
-                collection = []
-                for a_i in range(num_anim):
-                    if ~np.isnan(null_mean_array[a_i][c_i]) * ~np.isnan(null_std_array[a_i][c_i]):
-                        collection.extend(list(np.random.normal(null_mean_array[a_i][c_i],\
-                                                                null_std_array[a_i][c_i],\
-                                                                    50)))
-                new_null_mean_array[c_i] = np.nanmean(collection)
-                new_null_std_array[c_i] = np.nanstd(collection)
-            true_mean_array = np.nanmean(true_array,0)
-            true_std_array = np.nanstd(true_array,0)
-            #Plot normal
-            ax[s_i].plot(cutoff_array,true_mean_array,color=colors[0],label='True')
-            true_min = true_mean_array-true_std_array
-            true_min[true_min < 0] = 0
-            true_max = true_mean_array+true_std_array
-            ax[s_i].fill_between(cutoff_array,true_max,\
-                                 true_min,color=colors[0],\
-                                     alpha=0.3,label='True Std')
-            ax[s_i].plot(cutoff_array,new_null_mean_array,color=colors[1],label='Null')
-            null_min = new_null_mean_array-new_null_std_array
-            null_min[null_min < 0] = 0
-            null_max = new_null_mean_array+new_null_std_array
-            ax[s_i].fill_between(cutoff_array,null_min,\
-                                 null_max,color=colors[1],\
-                                     alpha=0.3,label='Null Std')
-            ax[s_i].legend(loc='upper right')
-            ax[s_i].set_xlabel(dev_null_stat + ' cutoffs')
-            ax[s_i].set_ylabel('Count')
-            ax[s_i].set_title(seg_name)
-            #Plot log scale
-            ax_log[s_i].plot(cutoff_array,np.log(true_mean_array),color=colors[0],label='True')
-            log_min = np.log(true_min)
-            log_max = np.log(true_max)
-            ax_log[s_i].fill_between(cutoff_array,np.log(true_min),\
-                                 np.log(true_max),color=colors[0],\
-                                     alpha=0.3,label='True Std')
-            ax_log[s_i].plot(cutoff_array,np.log(new_null_mean_array),color=colors[1],label='Null')
-            ax_log[s_i].fill_between(cutoff_array,np.log(null_min),\
-                                 np.log(null_max),color=colors[1],\
-                                     alpha=0.3,label='Null Std')
-            ax_log[s_i].legend(loc='upper right')
-            ax_log[s_i].set_xlabel(dev_null_stat + ' cutoffs')
-            ax_log[s_i].set_ylabel('Log(Count)')
-            ax_log[s_i].set_title(seg_name)
-        for s_i in range(len(unique_segment_names)):
-            ax[s_i].set_ylim([-1/10*max_val,max_val+1/10*max_val])
-            ax_log[s_i].set_ylim([0,np.log(max_val+1/10*max_val)])
-        f.suptitle(dev_null_stat)
-        plt.tight_layout()
-        f.savefig(os.path.join(results_dir,dev_null_stat+'_combined.png'))
-        f.savefig(os.path.join(results_dir,dev_null_stat+'_combined.svg'))
-        plt.close(f)
-        f_log.suptitle(dev_null_stat)
-        plt.tight_layout()
-        f_log.savefig(os.path.join(results_dir,dev_null_stat+'_combined_log.png'))
-        f_log.savefig(os.path.join(results_dir,dev_null_stat+'_combined_log.svg'))
-        plt.close(f_log)
     
 def cross_dataset_cp_plots(cp_data, unique_given_names, unique_taste_names, 
                            max_cp_counts, results_dir):
